requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.20.3=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
arviz=0.12.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
bitsandbytes=0.39.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
cchardet=2.1.7=pypi_0
certifi=2023.5.7=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
cfgv=3.3.1=pypi_0
cftime=1.6.2=pypi_0
chardet=5.1.0=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py38h9ce1e76_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pypi_0
datasets=2.13.0=pypi_0
decorator=5.1.1=pypi_0
deprecat=2.1.1=pypi_0
dill=0.3.6=pypi_0
distlib=0.3.6=pypi_0
evaluate=0.4.0=pypi_0
factor-analyzer=0.4.1=pypi_0
fastprogress=1.0.3=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.12.0=pypi_0
fonttools=4.39.3=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2023.6.0=pypi_0
ftfy=6.1.1=pypi_0
gensim=4.3.1=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.15.1=pypi_0
identify=2.5.24=pypi_0
idna=3.4=py38h06a4308_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py38h06a4308_0
joblib=1.2.0=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.7.1=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.6=py38h417a72b_1
mkl_random=1.2.2=py38h417a72b_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py38h06a4308_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
ncurses=6.4=h6a678d5_0
netcdf4=1.6.3=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py38h06a4308_1
nltk=3.8.1=pypi_0
nodeenv=1.7.0=pypi_0
numpy=1.22.4=pypi_0
objaverse=0.0.7=pypi_0
openai=0.27.2=pypi_0
openai-clip=1.0.1=pypi_0
opencv-python=4.7.0.72=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=pypi_0
pandas=2.0.0=pypi_0
patsy=0.5.3=pypi_0
pillow=9.4.0=pypi_0
pip=23.0.1=py38h06a4308_0
platformdirs=3.5.1=pypi_0
pre-commit=3.3.1=pypi_0
protobuf=4.23.3=pypi_0
psutil=5.9.5=pypi_0
py=1.11.0=pypi_0
pyarrow=12.0.1=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyglet=1.5.27=pypi_0
pymc3=3.11.5=pypi_0
pyopenssl=23.0.0=py38h06a4308_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pypi_0
pytorch=2.0.1=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
regex=2023.3.23=pypi_0
requests=2.28.2=pypi_0
responses=0.18.0=pypi_0
retry=0.9.2=pypi_0
safetensors=0.3.1=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.12.2=pypi_0
semver=3.0.0=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=65.6.3=py38h06a4308_0
six=1.16.0=pypi_0
smart-open=6.3.0=pypi_0
sqlite=3.41.1=h5eee18b_0
statsmodels=0.14.0=pypi_0
sympy=1.11.1=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
theano-pymc=1.1.2=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
torchaudio=2.0.2=py38_cu118
torchtriton=2.0.0=py38
torchvision=0.15.2=py38_cu118
tqdm=4.65.0=pypi_0
transformers=4.30.1=pypi_0
trimesh=3.21.5=pypi_0
typing_extensions=4.5.0=py38h06a4308_0
tzdata=2023.3=pypi_0
urllib3=1.26.15=py38h06a4308_0
virtualenv=20.23.0=pypi_0
wcwidth=0.2.6=pypi_0
wheel=0.38.4=py38h06a4308_0
wrapt=1.15.0=pypi_0
xarray=2023.1.0=pypi_0
xarray-einstats=0.5.1=pypi_0
xxhash=3.2.0=pypi_0
xz=5.2.10=h5eee18b_1
yarl=1.8.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
